# Toy.OpenApi/wwwroot/swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sample API",
    "description": "An ASP.NET Core Web API for managing sample items",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Example Contact",
      "url": "https://example.com/contact"
    },
    "license": {
      "name": "Example License",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5496"
    }
  ],
  "paths": {
    "/Samples": {
      "get": {
        "tags": [
          "Sample"
        ],
        "summary": "Creates a new product",
        "description": "...",
        "operationId": "CreateProduct",
        "parameters": [
          {
            "name": "Name",
            "in": "query",
            "description": "Search name keywords",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "From",
            "in": "query",
            "description": "From 조회",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "description": "To 조회",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The product was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sample"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sample"
        ],
        "summary": "Creates a new sample",
        "description": "...",
        "operationId": "CreateProduct",
        "parameters": [
          {
            "name": "Name",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The product was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sample"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Sample": {
        "required": [
          "샘플"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "샘플 아이디",
            "format": "int32",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "샘플 이름",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Sample",
      "description": "Create, read, update and delete sample"
    }
  ]
}